+++ apps/front/src/components/search/Courses/index.tsx
@@ -1,0 +1,8 @@
+import { Pagination } from './Pagination';
+import { Card, Cards } from './Card';
+
+export default {
+   Pagination,
+   Cards,
+   Card,
+};
