--- apps/front/src/components/search/index.ts
@@ -1,2 +1,0 @@
-export * from "./RefinementCheckbox";
-export * from "./RefinementSelect";
