+++ apps/front/src/components/search/refinements/index.tsx
@@ -1,0 +1,4 @@
+export * from './RefinementCheckbox';
+export * from './SortSelect';
+export * from './SearchBox';
+export * from './RefinementSelect';
