+++ apps/front/src/routeTree.gen.ts
@@ -14,7 +14,6 @@ import { Route as SearchRouteImport } from './routes/_search'
 import { Route as IndexRouteImport } from './routes/index'
 import { Route as SalaryDialogRouteImport } from './routes/salary/_dialog'
 import { Route as SearchCoursesRouteImport } from './routes/_search/courses'
-import { Route as SearchCoursesIndexRouteImport } from './routes/_search/courses/index'
 import { Route as SalaryDialogSubmissionsRouteImport } from './routes/salary/_dialog/submissions'
 import { Route as SalaryDialogDraftsRouteImport } from './routes/salary/_dialog/drafts'
 import { Route as SalaryDialogAutoFillRouteImport } from './routes/salary/_dialog/auto-fill'
@@ -44,11 +43,6 @@ const SearchCoursesRoute = SearchCoursesRouteImport.update({
   id: '/courses',
   path: '/courses',
   getParentRoute: () => SearchRoute,
-} as any)
-const SearchCoursesIndexRoute = SearchCoursesIndexRouteImport.update({
-  id: '/',
-  path: '/',
-  getParentRoute: () => SearchCoursesRoute,
 } as any)
 const SalaryDialogSubmissionsRoute = SalaryDialogSubmissionsRouteImport.update({
   id: '/submissions',
@@ -85,21 +79,20 @@ const SalaryDialogFormReportChar123IdxChar125Route =
 export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
   '/salary': typeof SalaryDialogFormRouteWithChildren
-  '/courses': typeof SearchCoursesRouteWithChildren
+  '/courses': typeof SearchCoursesRoute
   '/salary/auto-fill': typeof SalaryDialogAutoFillRoute
   '/salary/drafts': typeof SalaryDialogDraftsRoute
   '/salary/submissions': typeof SalaryDialogSubmissionsRoute
-  '/courses/': typeof SearchCoursesIndexRoute
   '/salary/report/{-$idx}': typeof SalaryDialogFormReportChar123IdxChar125Route
   '/salary/reported/$key': typeof SalaryDialogFormReportedKeyRoute
 }
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
   '/salary': typeof SalaryDialogFormRouteWithChildren
+  '/courses': typeof SearchCoursesRoute
   '/salary/auto-fill': typeof SalaryDialogAutoFillRoute
   '/salary/drafts': typeof SalaryDialogDraftsRoute
   '/salary/submissions': typeof SalaryDialogSubmissionsRoute
-  '/courses': typeof SearchCoursesIndexRoute
   '/salary/report/{-$idx}': typeof SalaryDialogFormReportChar123IdxChar125Route
   '/salary/reported/$key': typeof SalaryDialogFormReportedKeyRoute
 }
@@ -108,13 +101,12 @@ export interface FileRoutesById {
   '/': typeof IndexRoute
   '/_search': typeof SearchRouteWithChildren
   '/salary': typeof SalaryRouteWithChildren
-  '/_search/courses': typeof SearchCoursesRouteWithChildren
+  '/_search/courses': typeof SearchCoursesRoute
   '/salary/_dialog': typeof SalaryDialogRouteWithChildren
   '/salary/_dialog/_form': typeof SalaryDialogFormRouteWithChildren
   '/salary/_dialog/auto-fill': typeof SalaryDialogAutoFillRoute
   '/salary/_dialog/drafts': typeof SalaryDialogDraftsRoute
   '/salary/_dialog/submissions': typeof SalaryDialogSubmissionsRoute
-  '/_search/courses/': typeof SearchCoursesIndexRoute
   '/salary/_dialog/_form/report/{-$idx}': typeof SalaryDialogFormReportChar123IdxChar125Route
   '/salary/_dialog/_form/reported/$key': typeof SalaryDialogFormReportedKeyRoute
 }
@@ -127,17 +119,16 @@ export interface FileRouteTypes {
     | '/salary/auto-fill'
     | '/salary/drafts'
     | '/salary/submissions'
-    | '/courses/'
     | '/salary/report/{-$idx}'
     | '/salary/reported/$key'
   fileRoutesByTo: FileRoutesByTo
   to:
     | '/'
     | '/salary'
+    | '/courses'
     | '/salary/auto-fill'
     | '/salary/drafts'
     | '/salary/submissions'
-    | '/courses'
     | '/salary/report/{-$idx}'
     | '/salary/reported/$key'
   id:
@@ -151,7 +142,6 @@ export interface FileRouteTypes {
     | '/salary/_dialog/auto-fill'
     | '/salary/_dialog/drafts'
     | '/salary/_dialog/submissions'
-    | '/_search/courses/'
     | '/salary/_dialog/_form/report/{-$idx}'
     | '/salary/_dialog/_form/reported/$key'
   fileRoutesById: FileRoutesById
@@ -199,13 +189,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof SearchCoursesRouteImport
       parentRoute: typeof SearchRoute
     }
-    '/_search/courses/': {
-      id: '/_search/courses/'
-      path: '/'
-      fullPath: '/courses/'
-      preLoaderRoute: typeof SearchCoursesIndexRouteImport
-      parentRoute: typeof SearchCoursesRoute
-    }
     '/salary/_dialog/submissions': {
       id: '/salary/_dialog/submissions'
       path: '/submissions'
@@ -251,24 +234,12 @@ declare module '@tanstack/react-router' {
   }
 }
 
-interface SearchCoursesRouteChildren {
-  SearchCoursesIndexRoute: typeof SearchCoursesIndexRoute
-}
-
-const SearchCoursesRouteChildren: SearchCoursesRouteChildren = {
-  SearchCoursesIndexRoute: SearchCoursesIndexRoute,
-}
-
-const SearchCoursesRouteWithChildren = SearchCoursesRoute._addFileChildren(
-  SearchCoursesRouteChildren,
-)
-
 interface SearchRouteChildren {
-  SearchCoursesRoute: typeof SearchCoursesRouteWithChildren
+  SearchCoursesRoute: typeof SearchCoursesRoute
 }
 
 const SearchRouteChildren: SearchRouteChildren = {
-  SearchCoursesRoute: SearchCoursesRouteWithChildren,
+  SearchCoursesRoute: SearchCoursesRoute,
 }
 
 const SearchRouteWithChildren =
